+++ adminer/sql.inc.php
@@ -235,7 +235,9 @@ if (!isset($_GET["import"])) {
 	echo "<p>";
 	textarea("query", $q, 20);
 	echo script(($_POST ? "" : "qs('textarea').focus();\n") . "qs('#form').onsubmit = partial(sqlSubmit, qs('#form'), '" . js_escape(remove_from_uri("sql|limit|error_stops|only_errors|history")) . "');");
-	echo "<p>$execute\n";
+	echo "<p>";
+	$adminer->sqlPrintAfter();
+	echo "$execute\n";
 	echo lang('Limit rows') . ": <input type='number' name='limit' class='size' value='" . h($_POST ? $_POST["limit"] : $_GET["limit"]) . "'>\n";
 
 } else {