+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Plugins: autoload plugins from adminer-plugins/
 - Plugins: configure plugins with adminer-plugins.php
 - Plugins: Display loaded plugins in server overview
+- New plugin: AI prompt in SQL command generating the queries with Google Gemini
 - New plugin: IMAP driver created for fun
 - New plugin: Display links to tables referencing current row
 - New plugin: Allow switching light and dark mode (bug #926)